+++ src/clients/IOClients.ts
@@ -2,7 +2,7 @@ import { InstanceOptions } from '../HttpClient'
 import { IOContext } from '../service/worker/runtime/typings'
 import { Billing, Builder, MessagesGraphQL, Settings } from './apps'
 import { CatalogGraphQL } from './apps/catalogGraphQL/index'
-import { ID, PaymentProvider } from './external'
+import { ID, MasterData, PaymentProvider } from './external'
 import {
   Apps,
   Assets,
@@ -60,6 +60,10 @@ export class IOClients {
 
   public get licenseManager() {
     return this.getOrSet('licenseManager', LicenseManager)
+  }
+
+  public get masterdata() {
+    return this.getOrSet('masterdata', MasterData)
   }
 
   public get messagesGraphQL() {